Use lower letters for charset (`utf-8`)
<https://html.spec.whatwg.org/multipage/semantics.html#attr-meta-charset>
says:

> The charset attribute specifies the character encoding used by the
document. This is a character encoding declaration. If the attribute is
present, its value must be an ASCII case-insensitive match for the
string "utf-8".

This is case-insensitive so there are no harm with `UTF-8`, but I'd like
to use the exactly same string as the example on the spec.